refractoring of code to contain less duplication
